RGB values for colors are now represented using double

+++ src/Protocol/Models/Color.cs
@@ -8,21 +8,21 @@ namespace OmniSharp.Extensions.LanguageServer.Protocol.Models
         /// <summary>
         /// The red component of this color in the range [0-1].
         /// </summary>
-        public int Red { get; set; }
+        public double Red { get; set; }
 
         /// <summary>
         /// The green component of this color in the range [0-1].
         /// </summary>
-        public int Green { get; set; }
+        public double Green { get; set; }
 
         /// <summary>
         /// The blue component of this color in the range [0-1].
         /// </summary>
-        public int Blue { get; set; }
+        public double Blue { get; set; }
 
         /// <summary>
         /// The alpha component of this color in the range [0-1].
         /// </summary>
-        public int Alpha { get; set; }
+        public double Alpha { get; set; }
     }
 }
